refactor ability and ai scaffolds
