+++ .sampo/changesets/673-modularize-ability-ai-scaffolds.md
@@ -1,0 +1,8 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Modularize the workspace `add ability` and `add ai-feature` scaffolds into
+thin entry points plus focused template and workspace-mutation helpers while
+keeping representative generated output and integration behavior stable.
